requirements.txt
django==3.1.1
djangorestframework>=3.11.1,<3.12.0
django-filter==2.3.0
djangorestframework-jwt>=1.11.0,<1.12.0
django-jazzmin>=2.2.11,<2.3.0
django-import-export>=2.4.0,<2.5.0
django-resized>=0.3.11,<0.4.0
django-cors-headers>=3.5.0,<3.6.0
django-credit-cards >= 0.4.1
django-ckeditor>=6.0.0, <6.1.0
django-elasticsearch-dsl-drf >= 0.21
django-modeltranslation >= 0.16.2

psycopg2>=2.8.6,<2.9.0
Pillow>=7.2.0,<7.3.0
drf-yasg>=1.17.1,<1.18.0
cryptography == 2.8
flake8>=3.6.0,<3.7.0
boto3 == 1.17.1
django-storages == 1.11.1



